+++ python/tvm/relay/transform/memory_alloc.py
@@ -42,8 +42,8 @@ class ManifestAllocPass(ExprMutator):
     """A pass for explicitly manifesting all memory allocations in Relay."""
 
     def __init__(self, target_host):
-        self.invoke_tvm = op.memory.invoke_tvm_op
-        self.shape_func = op.memory.shape_func
+        self.invoke_tvm = op.vm.invoke_tvm_op
+        self.shape_func = op.vm.shape_func
         self.shape_of = op.vm.shape_of
         self.scopes = [ScopeBuilder()]
         self.target_host = target_host
